serwer/requirements.txt
asgiref==3.3.4
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
dateparser==1.0.0
Django==3.2
djangorestframework==3.12.4
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
numpy==1.20.2
pandas==1.2.4
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pymongo==3.11.3
pyOpenSSL==20.0.1
python-binance==0.7.10
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.4
zope.interface==5.4.0

coinpaprika~=0.1.0
